pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finahunt"
version = "0.1.0"
description = "Compliance information source access agent system scaffold"
requires-python = ">=3.11"
dependencies = [
  "langgraph>=0.2.60",
  "pydantic>=2.8.2",
  "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "packages*", "graphs*", "workflows*", "skills*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
